Address more review comments.
